# Fix ZuluIDE properly connecting to the Pico W