--- a/.gitignore
+++ b/.gitignore
@@ -1,0 +1,2 @@
+build/
+src/index_html.h