The main issue was the reset watchdog was called when the ZuluIDE was initializing took just too long in some cases. Changing it from 1sec to 10ms fixed most of the issues with different SD cards. It also allowed the OLED controller board to be connected with HTTP PicoW to the ZuluIDE.

Also fixed was the GPIO pin number for SDA and SCL which were flipped. It had no real affect on anything as their initializing didn't care on the pin number being swapped.

Lastly edited the linker script to put most of the compiled code in the src directory in SRAM. It seemed to speed up the ZuluIDE from 9MB/s to 9.1MB/s. A small increase that can be reversed SRAM usage become much larger.